+++ coaster/sqlalchemy/statemanager.py
@@ -200,13 +200,12 @@ from collections import OrderedDict
 import functools
 from sqlalchemy import and_, or_, column as column_constructor, CheckConstraint
 from werkzeug.exceptions import BadRequest
-from ..utils import NameTitle
+from ..utils import is_collection, NameTitle
 from ..signals import coaster_signals
+from .roles import RoleMixin
 
 __all__ = ['StateManager', 'StateTransitionError',
     'transition_error', 'transition_before', 'transition_after', 'transition_exception']
-
-iterables = (set, frozenset, list, tuple)  # Used for various isinstance checks
 
 
 # --- Signals -----------------------------------------------------------------
@@ -256,12 +255,12 @@ class ManagedState(object):
     @property
     def is_scalar(self):
         """This is a scalar state (not a group of states)"""
-        return not isinstance(self.value, iterables)
+        return not is_collection(self.value)
 
     @property
     def is_direct(self):
         """This is a direct state (scalar state without a condition)"""
-        return self.validator is None and not isinstance(self.value, iterables)
+        return self.validator is None and not is_collection(self.value)
 
     def __repr__(self):
         return '%s.%s' % (self.statemanager.name, self.name)
@@ -269,7 +268,7 @@ class ManagedState(object):
     def _eval(self, obj, cls=None):
         # TODO: Respect cache as specified in `cache_for`
         if obj is not None:  # We're being called with an instance
-            if isinstance(self.value, iterables):
+            if is_collection(self.value):
                 valuematch = self.statemanager._value(obj, cls) in self.value
             else:
                 valuematch = self.statemanager._value(obj, cls) == self.value
@@ -278,7 +277,7 @@ class ManagedState(object):
             else:
                 return valuematch
         else:  # We have a class, so return a filter condition, for use as cls.query.filter(result)
-            if isinstance(self.value, iterables):
+            if is_collection(self.value):
                 valuematch = self.statemanager._value(obj, cls).in_(self.value)
             else:
                 valuematch = self.statemanager._value(obj, cls) == self.value
@@ -320,7 +319,7 @@ class ManagedStateGroup(object):
         values = set()
         for state in regular_states:
             self.states.append(state)
-            if isinstance(state.value, iterables):
+            if is_collection(state.value):
                 values.update(state.value)
             else:
                 values.add(state.value)
@@ -329,7 +328,7 @@ class ManagedStateGroup(object):
         # regular state. This is an error as the condition will never be tested
         for state in conditional_states:
             # Prevent grouping of conditional states with their original states
-            state_values = set(state.value if isinstance(state.value, iterables) else [state.value])
+            state_values = set(state.value if is_collection(state.value) else [state.value])
             if state_values & values:  # They overlap
                 raise ValueError("The value for state %s is already in this state group" % repr(state))
             self.states.append(state)
@@ -413,16 +412,15 @@ class StateTransition(object):
             raise StateTransitionError("Duplicate transition decorator")
         if from_ is not None and not isinstance(from_, (ManagedState, ManagedStateGroup)):
             raise StateTransitionError("From state is not a managed state: %s" % repr(from_))
-        if not isinstance(to, ManagedState):
-            raise StateTransitionError("To state is not a managed state: %s" % repr(to))
         if from_ and from_.statemanager != statemanager:
             raise StateTransitionError("From state is not managed by this state manager: %s" % repr(from_))
-        if to.statemanager != statemanager:
-            raise StateTransitionError("To state is not managed by this state manager: %s" % repr(to))
-        elif to.validator is not None:
-            raise StateTransitionError("To state cannot be a conditional state: %s" % repr(to))
-        elif isinstance(to.value, iterables) or to.value not in statemanager.lenum:
-            raise StateTransitionError("To state's value is not a valid value: %s" % repr(to))
+        if to is not None:
+            if not isinstance(to, ManagedState):
+                raise StateTransitionError("To state is not a managed state: %s" % repr(to))
+            if to.statemanager != statemanager:
+                raise StateTransitionError("To state is not managed by this state manager: %s" % repr(to))
+            if not to.is_direct:
+                raise StateTransitionError("To state must be a direct state: %s" % repr(to))
         if data:
             if 'name' in data:
                 raise TypeError("Invalid transition data parameter 'name'")
@@ -446,15 +444,15 @@ class StateTransition(object):
                 from_ = [from_]
             # Step 2: Unroll grouped values from the original LabeledEnum
             for mstate in from_:
-                if isinstance(mstate.value, iterables):
+                if is_collection(mstate.value):
                     for value in mstate.value:
                         state_values[value] = mstate
                 else:
                     state_values[mstate.value] = mstate
 
         self.transitions[statemanager] = {
-            'from': state_values,  # Just the valuesScalar values (no validation functions)
-            'to': to,              # Scalar value of new state
+            'from': state_values,  # Dict of scalar_value: ManagedState
+            'to': to,              # ManagedState (is_direct) of new state
             'if': if_,             # Additional conditions that must ALL pass
             }
 
@@ -527,7 +525,8 @@ class StateTransitionWrapper(object):
             raise
         # Change the state for each of the state managers
         for statemanager, conditions in self.statetransition.transitions.items():
-            statemanager._set(self.obj, conditions['to'].value)  # Change state
+            if conditions['to'] is not None:  # Allow to=None for the @requires decorator
+                statemanager._set(self.obj, conditions['to'].value)  # Change state
         # Raise a transition-after signal
         transition_after.send(self.obj, transition=self.statetransition)
         return result
@@ -678,7 +677,7 @@ class StateManager(object):
         If it returns without an error, the state value is updated
         automatically.
 
-        :param from_: Required original state to allow this transition (can be a group of states)
+        :param from_: Required state to allow this transition (can be a state group)
         :param to: The state of the object after this transition (automatically set if no exception is raised)
         :param if_: Validator(s) that, given the object, must all return True for the transition to proceed
         :param metadata: Additional metadata, stored on the StateTransition object
@@ -693,6 +692,17 @@ class StateManager(object):
             return st
 
         return decorator
+
+    def requires(self, from_, if_=None, **data):
+        """
+        Decorates a method that may be called if the given state is currently active.
+        Registers a transition internally, but does not change the state.
+
+        :param from_: Required state to allow this call (can be a state group)
+        :param if_: Validator(s) that, given the object, must all return True for the call to proceed
+        :param metadata: Additional metadata, stored on the StateTransition object
+        """
+        return self.transition(from_, None, if_, **data)
 
     def _value(self, obj, cls=None):
         """The state value (called from the wrapper)"""
@@ -765,15 +775,31 @@ class StateManagerWrapper(object):
                 for name, mstate in self.statemanager.states.items()
                 if mstate(self.obj, self.cls)}
 
-    @property
-    def transitions(self):
+    def transitions(self, current=True):
         """
-        Returns currently available transitions as a dictionary of name: StateTransitionWrapper
+        Returns available transitions for the current state, as a dictionary of
+        name: StateTransitionWrapper.
+
+        :param bool current: Limit to transitions available in ``obj.current_access()``
         """
-        # Retrieve transitions from the instance object to activate the descriptor.
+        if current and isinstance(self.obj, RoleMixin):
+            proxy = self.obj.current_access()
+        else:
+            proxy = {}
+            current = False  # In case the host object is not a RoleMixin
         return {name: transition for name, transition in
+            # Retrieve transitions from the host object to activate the descriptor.
             ((name, getattr(self.obj, name)) for name in self.statemanager.transitions)
-            if transition.is_available}
+            if transition.is_available and (name in proxy if current else True)}
+
+    def transitions_for(self, roles=None, actor=None, anchors=[]):
+        """
+        For use on RoleMixin classes: returns currently available transitions for the specified
+        roles or actor as a dictionary of name: StateTransitionWrapper.
+        """
+        proxy = self.obj.access_for(roles, actor, anchors)
+        return {name: transition for name, transition in self.transitions(current=False).items()
+            if name in proxy}
 
     def group(self, items, keep_empty=False):
         """